+++ data_safe_haven/config/__init__.py
@@ -1,6 +1,6 @@
 from .config import Config
-from .pulumi import DSHPulumiConfig
-from .pulumi_project import DSHPulumiProject
+from .dsh_pulumi_config import DSHPulumiConfig
+from .dsh_pulumi_project import DSHPulumiProject
 
 __all__ = [
     "Config",
